+++ dev-tools/mage/gotest.go
@@ -83,7 +83,11 @@ func makeGoTestArgsForModule(name, module string) GoTestArgs {
 // testTagsFromEnv gets a list of comma-separated tags from the TEST_TAGS
 // environment variables, e.g: TEST_TAGS=aws,azure.
 func testTagsFromEnv() []string {
-	return strings.Split(strings.Trim(os.Getenv("TEST_TAGS"), ", "), ",")
+	tags := strings.Split(strings.Trim(os.Getenv("TEST_TAGS"), ", "), ",")
+	if FIPSBuild {
+		tags = append(tags, "requirefips")
+	}
+	return tags
 }
 
 // DefaultGoTestUnitArgs returns a default set of arguments for running
