+++ internal/pkg/agent/application/actions/handlers/handler_action_policy_change_test.go
@@ -6,10 +6,8 @@ package handlers
 
 import (
 	"context"
-	"crypto/rand"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
@@ -498,21 +496,6 @@ func TestPolicyChangeHandler_handlePolicyChange_FleetClientSettings(t *testing.T
 		agentRootPair, agentChildPair, err := certutil.NewRootAndChildCerts()
 		require.NoError(t, err, "failed creating root and child certs")
 
-		agentChildDERKey, _ := pem.Decode(agentChildPair.Key)
-		require.NoError(t, err, "could not create tls.Certificates from child certificate")
-
-		encPem, err := x509.EncryptPEMBlock( //nolint:staticcheck // we need to drop support for this, but while we don't, it needs to be tested.
-			rand.Reader,
-			"EC PRIVATE KEY",
-			agentChildDERKey.Bytes,
-			[]byte(agentChildEncPassphrase),
-			x509.PEMCipherAES128)
-		require.NoError(t, err, "failed encrypting agent child certificate key block")
-		agentChildEncPair := certutil.Pair{
-			Cert: agentChildPair.Cert,
-			Key:  pem.EncodeToMemory(encPem),
-		}
-
 		wrongRootPair, wrongChildPair, err := certutil.NewRootAndChildCerts()
 		require.NoError(t, err, "failed creating root and child certs")
 
@@ -708,40 +691,6 @@ func TestPolicyChangeHandler_handlePolicyChange_FleetClientSettings(t *testing.T
 				},
 			},
 			{
-				name: "certificate and key with passphrase is applied when present",
-				originalCfg: &configuration.Configuration{
-					Fleet: &configuration.FleetAgentConfig{
-						Client: remote.Config{
-							Host: fleetmTLSServer.URL,
-							Transport: httpcommon.HTTPTransportSettings{
-								TLS: &tlscommon.Config{
-									CAs: []string{string(fleetRootPair.Cert)},
-								},
-							},
-						},
-						AccessAPIKey: "ignore",
-					},
-					Settings: configuration.DefaultSettingsConfig(),
-				},
-				newCfg: map[string]interface{}{
-					"fleet.ssl.enabled":        true,
-					"fleet.ssl.certificate":    string(agentChildEncPair.Cert),
-					"fleet.ssl.key":            string(agentChildEncPair.Key),
-					"fleet.ssl.key_passphrase": agentChildEncPassphrase,
-				},
-				setterCalledCount: 1,
-				wantCAs:           []string{string(fleetRootPair.Cert)},
-				wantCertificateConfig: tlscommon.CertificateConfig{
-					Certificate: string(agentChildEncPair.Cert),
-					Key:         string(agentChildEncPair.Key),
-					Passphrase:  agentChildEncPassphrase,
-				},
-				assertErr: func(t *testing.T, err error) {
-					assert.NoError(t, err,
-						"unexpected error when applying fleet.ssl.certificate and key")
-				},
-			},
-			{
 				name: "certificate and key without passphrase clear out previous passphrase",
 				originalCfg: &configuration.Configuration{
 					Fleet: &configuration.FleetAgentConfig{
@@ -775,40 +724,6 @@ func TestPolicyChangeHandler_handlePolicyChange_FleetClientSettings(t *testing.T
 					Key:            string(agentChildPair.Key),
 					Passphrase:     "",
 					PassphrasePath: "",
-				},
-				assertErr: func(t *testing.T, err error) {
-					assert.NoError(t, err,
-						"unexpected error when applying fleet.ssl.certificate and key")
-				},
-			},
-			{
-				name: "certificate and key with passphrase_path is applied when present",
-				originalCfg: &configuration.Configuration{
-					Fleet: &configuration.FleetAgentConfig{
-						Client: remote.Config{
-							Host: fleetmTLSServer.URL,
-							Transport: httpcommon.HTTPTransportSettings{
-								TLS: &tlscommon.Config{
-									CAs: []string{string(fleetRootPair.Cert)},
-								},
-							},
-						},
-						AccessAPIKey: "ignore",
-					},
-					Settings: configuration.DefaultSettingsConfig(),
-				},
-				newCfg: map[string]interface{}{
-					"fleet.ssl.enabled":             true,
-					"fleet.ssl.certificate":         string(agentChildEncPair.Cert),
-					"fleet.ssl.key":                 string(agentChildEncPair.Key),
-					"fleet.ssl.key_passphrase_path": passphrasePath,
-				},
-				setterCalledCount: 1,
-				wantCAs:           []string{string(fleetRootPair.Cert)},
-				wantCertificateConfig: tlscommon.CertificateConfig{
-					Certificate:    string(agentChildEncPair.Cert),
-					Key:            string(agentChildEncPair.Key),
-					PassphrasePath: passphrasePath,
 				},
 				assertErr: func(t *testing.T, err error) {
 					assert.NoError(t, err,
